+++ airbyte-cdk/python/airbyte_cdk/sources/declarative/manifest_declarative_source.py
@@ -118,10 +118,6 @@ class ManifestDeclarativeSource(DeclarativeSource):
         self._configure_logger_level(logger)
         yield from super().read(logger, config, catalog, state)
 
-    def log_slice_message(self, logger: logging.Logger):
-        # Commented so the unit tests fail
-        return self._constructor._emit_connector_builder_messages or super(self).log_slice_message(logger)
-
     def _configure_logger_level(self, logger: logging.Logger):
         """
         Set the log level to logging.DEBUG if debug mode is enabled
